+++ frontend/src/app/components/containers/complaints/list-items/allegation-complaint-list-item.tsx
@@ -15,6 +15,7 @@ import { getUserAgency } from "@/app/service/user-service";
 import { FeatureFlag } from "@/app/components/common/feature-flag";
 import { FEATURE_TYPES } from "@/app/constants/feature-flag-types";
 import { selectParkByGuid } from "@/app/store/reducers/park";
+import { usePark } from "@/app/hooks/usePark";
 
 type Props = {
   type: string;
@@ -46,7 +47,7 @@ export const AllegationComplaintListItem: FC<Props> = ({ type, complaint }) => {
 
   const userAgency = getUserAgency();
   const derivedAllegationStatus = ownedBy !== userAgency ? "Referred" : status;
-  const park = useAppSelector(selectParkByGuid(parkGuid));
+  const park = usePark(parkGuid);
   const parkAreaGuids = park?.parkAreas?.map((area) => area.parkAreaGuid) ?? [];
 
   const getStatusDescription = (input: string): string => {
