+++ frontend/src/app/components/containers/complaints/list-items/general-complaint-list-item.tsx
@@ -11,6 +11,7 @@ import { getUserAgency } from "@/app/service/user-service";
 import { FeatureFlag } from "@/app/components/common/feature-flag";
 import { FEATURE_TYPES } from "@/app/constants/feature-flag-types";
 import { selectParkByGuid } from "@/app/store/reducers/park";
+import { usePark } from "@/app/hooks/usePark";
 
 type Props = {
   type: string;
@@ -42,7 +43,7 @@ export const GeneralInformationComplaintListItem: FC<Props> = ({ type, complaint
 
   const userAgency = getUserAgency();
   const derivedGeneralStatus = ownedBy !== userAgency ? "Referred" : status;
-  const park = useAppSelector(selectParkByGuid(parkGuid));
+  const park = usePark(parkGuid);
   const parkAreaGuids = park?.parkAreas?.map((area) => area.parkAreaGuid) ?? [];
 
   const getLocationName = (input: string): string => {
