+++ frontend/src/app/components/containers/complaints/list-items/wildlife-complaint-list-item.tsx
@@ -12,6 +12,7 @@ import { getUserAgency } from "@/app/service/user-service";
 import { FeatureFlag } from "@/app/components/common/feature-flag";
 import { FEATURE_TYPES } from "@/app/constants/feature-flag-types";
 import { selectParkByGuid } from "@/app/store/reducers/park";
+import { usePark } from "@/app/hooks/usePark";
 
 type Props = {
   type: string;
@@ -44,7 +45,7 @@ export const WildlifeComplaintListItem: FC<Props> = ({ type, complaint }) => {
 
   const userAgency = getUserAgency();
   const derivedWildlifeStatus = ownedBy !== userAgency ? "Referred" : status;
-  const park = useAppSelector(selectParkByGuid(parkGuid));
+  const park = usePark(parkGuid);
   const parkAreaGuids = park?.parkAreas?.map((area) => area.parkAreaGuid) ?? [];
 
   const getStatusDescription = (input: string): string => {
