+++ frontend/src/app/store/reducers/park.ts
@@ -1,6 +1,9 @@
 import { Park } from "@/app/types/app/shared/park";
-import { createSelector, createSlice, PayloadAction } from "@reduxjs/toolkit";
+import { Action, createSelector, createSlice, PayloadAction, ThunkAction } from "@reduxjs/toolkit";
 import { RootState } from "../store";
+import { generateApiParameters, get } from "@/app/common/api";
+import config from "@/config";
+import { UUID } from "crypto";
 
 interface ParkState {
   parkCache: Record<string, Park>;
@@ -28,4 +31,37 @@ export const selectParkByGuid = (guid: string | undefined) =>
   createSelector([selectParkEntities], (parkCache) => (guid ? parkCache[guid] : undefined));
 
 export const { setPark } = parksReducer.actions;
+
+// get parkData for a single Park
+export const getParkData =
+  (parkId: string): ThunkAction<Promise<Park | undefined>, RootState, unknown, Action<string>> =>
+  async (dispatch, getState) => {
+    // check and see if we've got a cached copy of this park
+    const existingPark = getState().parks.parkCache[parkId];
+    if (existingPark && !existingPark.isFallback) {
+      return existingPark;
+    }
+    //If not fetch from the shared database
+    try {
+      const parameters = generateApiParameters(`${config.API_BASE_URL}/v1/shared-data/park/${parkId}`);
+      const response = await get<Park>(dispatch, parameters);
+
+      if (response) {
+        dispatch(setPark(response));
+        return response;
+      } else {
+        const fallback = {
+          name: "Park name unavailable",
+          parkGuid: parkId as UUID,
+          isFallback: true,
+        };
+        dispatch(setPark(fallback));
+        return fallback;
+      }
+    } catch (error) {
+      console.error(`Unable to retrieve park information for ${parkId}: ${error}`);
+      return undefined;
+    }
+  };
+
 export default parksReducer.reducer;
